Fix 500 intro'd in babade7: create record req's data.add_record perm
Introduced a permissions requirement in babade7, which checked
data.change_record, but that meant it failed some checks while bypassing
others. This 500'd with a reverse() failure when creating personal records
(which we should forbid for anon soon), so we also check data.change_record
and return a 403 instead of a 500 on error.

And, if we can't create/edit records, and we just created one anyways
(possibly because we can create, but not edit?) and we have no id/name
to our name, then we can't just display the orphan record. Instead,
we 403 without creating the record.